refactor: removed to file all google moves
